Revert "Webpack Encore upgrade (#274)"
This reverts commit 71707fcc13ee8ed488332ff2856c1025b04aceee.
