+++ .circleci/config.yml
@@ -10,7 +10,7 @@ jobs:
             - vendor-{{ checksum "composer.lock" }}
       - restore_cache:
           keys:
-            - node_modules-{{ checksum "yarn.lock" }}
+            - node_modules-{{ checksum "package-lock.json" }}
       - run:
           name: Build Dependencies
           command: |
@@ -38,17 +38,14 @@ jobs:
             # Removes boilerplate feature test added during Symfony Flex recipe for behat
             rm -rf features/demo.feature
 
-            # Install js dependencies
-            docker-compose run --rm yarn install
-
             # Generate static assets
-            docker-compose run --rm yarn build-dev
+            docker-compose run --rm npm
       - save_cache:
           key: vendor-{{ checksum "composer.lock" }}
           paths:
           - vendor
       - save_cache:
-          key: node_modules-{{ checksum "yarn.lock" }}
+          key: node_modules-{{ checksum "package-lock.json" }}
           paths:
             - node_modules
       - run:
@@ -81,10 +78,8 @@ jobs:
           command: |
             # Vendor php dependencies
             docker-compose run --rm composer
-            # install js dependencies
-            docker-compose run --rm yarn install
             # Generate static assets
-            docker-compose run --rm yarn build
+            docker-compose run --rm npm
             # Build app
             docker-compose build web
             docker-compose build app
