+++ .gitignore
@@ -11,7 +11,9 @@
 behat.yml
 ###< behat/symfony2-extension ###
 .idea
+# ignore bin as created by composer
 **/public/assets
+node_modules
 */bin/jp.php
 **/vendor
 bootstrap.php.cache
@@ -36,10 +38,3 @@ localstack-data
 
 /s3/
 /web/
-
-###> symfony/webpack-encore-bundle ###
-/node_modules/
-/public/build/
-npm-debug.log
-yarn-error.log
-###< symfony/webpack-encore-bundle ###
