+++ assets/javascript/modules/sessionTimeoutDialog.js
@@ -60,5 +60,3 @@ var SessionTimeoutDialog = function (options) {
     };
 
 };
-
-export default SessionTimeoutDialog
