--- assets/js/PageSpecific/addDeputy.js
@@ -1,7 +1,0 @@
-import PostcodeLookup from '../Components/postcodeLookup'
-
-window.addEventListener('DOMContentLoaded', (event) => {
-    let postcodeLookup = new PostcodeLookup();
-    postcodeLookup.cacheEls('.js-PostcodeLookup');
-    postcodeLookup.bindEvents();
-});
