--- assets/js/PageSpecific/orderSummary.js
@@ -1,9 +1,0 @@
-import InlineUpload from '../Components/inlineUpload'
-
-window.addEventListener('DOMContentLoaded', (event) => {
-    let inlineUpload = new InlineUpload();
-    inlineUpload.cacheEls();
-    inlineUpload.init();
-});
-
-
