--- assets/js/app.js
@@ -1,17 +1,0 @@
-/*
- * Welcome to your app's main JavaScript file!
- *
- * We recommend including the built version of this JavaScript file
- * (and its CSS file) in your base layout (base.html.twig).
- */
-
-// any CSS you require will output into a single css file (app.css in this case)
-require('../css/app.scss');
-require('dropzone/dist/min/dropzone.min');
-
-import { initAll } from 'govuk-frontend'
-
-initAll();
-
-// Need jQuery? Install it with "yarn add jquery", then uncomment to require it.
-const $ = require('jquery');
