+++ assets/scss/application.scss
@@ -1,9 +1,9 @@
 // Import GOVUK Frontend
 $govuk-global-styles: true;
 $govuk-page-width: 1200px;
-$govuk-assets-path: '/build/';
+$govuk-assets-path: '../';
 $govuk-font-family: helvetica, arial;
-@import "~govuk-frontend/all";
+@import "all";
 
 // GOVUK extensions
 @import "govuk-extension/core/links";
