+++ assets/scss/components/_alerts.scss
@@ -1,9 +1,10 @@
 // Alerts
 // ==========================================================================
 
-@import "~govuk-frontend/settings/all";
-@import "~govuk-frontend/tools/all";
-@import "~govuk-frontend/helpers/all";
+@import "settings/all";
+@import "tools/all";
+@import "helpers/all";
+
 
 @include govuk-exports("govuk/component/alert") {
     .govuk-alert {
