+++ assets/scss/components/_pagination.scss
@@ -1,9 +1,9 @@
 // Pagination
 // ==========================================================================
 
-@import "~govuk-frontend/settings/all";
-@import "~govuk-frontend/tools/all";
-@import "~govuk-frontend/helpers/all";
+@import "settings/all";
+@import "tools/all";
+@import "helpers/all";
 
 
 @include govuk-exports("govuk/component/pagination") {
