+++ assets/scss/govuk-extension/core/_links.scss
@@ -1,9 +1,9 @@
 // Links
 // ==========================================================================
 
-@import "~govuk-frontend/settings/all";
-@import "~govuk-frontend/tools/all";
-@import "~govuk-frontend/helpers/all";
+@import "settings/all";
+@import "tools/all";
+@import "helpers/all";
 
 
 @include govuk-exports("govuk/core/links-extended") {
