+++ composer.json
@@ -15,7 +15,7 @@
         "behat/mink-goutte-driver": "^1.2",
         "behat/symfony2-extension": "2.1.*",
         "doctrine/doctrine-bundle": "^1.9",
-        "doctrine/doctrine-fixtures-bundle": "^3.1",
+        "doctrine/doctrine-fixtures-bundle": "^3.0",
         "doctrine/doctrine-migrations-bundle": "^1.3",
         "doctrine/orm": "^2.6",
         "mockery/mockery": "1.0.*",
@@ -38,7 +38,6 @@
         "symfony/test-pack": "^1.0",
         "symfony/twig-bundle": "4.2.*",
         "symfony/validator": "4.2.*",
-        "symfony/webpack-encore-bundle": "^1.6",
         "symfony/yaml": "4.2.*",
         "twig/extensions": "1.5.*"
     },
