--- config/packages/assets.yaml
@@ -1,3 +1,0 @@
-framework:
-    assets:
-        json_manifest_path: '%kernel.project_dir%/public/build/manifest.json'
