--- config/packages/prod/webpack_encore.yaml
@@ -1,4 +1,0 @@
-#webpack_encore:
-    # Cache the entrypoints.json (rebuild Symfony's cache when entrypoints.json changes)
-    # Available in version 1.2
-    #cache: true
