+++ package.json
@@ -5,58 +5,21 @@
   "repository": "https://github.com/ministryofjustice/serve-opg",
   "license": "MIT",
   "devDependencies": {
-    "@babel/preset-env": "^7.4.5",
-    "@symfony/webpack-encore": "^0.27.0",
-    "babel-jest": "^24.8.0",
-    "core-js": "^3.0.0",
-    "del": "^3.0.0",
-    "dropzone": "5.5.1",
-    "govuk-frontend": "^2.12.0",
+    "govuk-frontend": "^2.7.0",
     "gulp": "^4.0.0",
     "gulp-clean": "^0.4.0",
     "gulp-concat": "^2.6.1",
-    "gulp-plumber": "^1.2.1",
     "gulp-sass": "^4.0.2",
-    "jest": "^24.8.0",
-    "jquery": "^3.4.1",
-    "sass-loader": "^7.0.1",
-    "tar": ">=4.4.2",
-    "webpack-notifier": "^1.6.0"
+    "jquery": "^3.4.0",
+    "del": "^3.0.0",
+    "gulp-plumber": "^1.2.1",
+    "dropzone": "5.5.1",
+    "tar": ">=4.4.2"
   },
   "scripts": {
-    "build": "encore production --progress",
-    "dev-server": "encore dev-server",
-    "build-dev": "encore dev",
-    "jest": "jest assets/js/__tests__",
-    "watch": "encore dev --watch",
-    "clear-jest-cache": "jest --clearCache"
-  },
-  "babel": {
-    "presets": [
-      [
-        "@babel/preset-env"
-      ]
-    ]
-  },
-  "jest": {
-    "moduleFileExtensions": [
-      "js"
-    ],
-    "moduleDirectories": [
-      "node_modules",
-      "src"
-    ],
-    "restoreMocks": true,
-    "testRegex": "\\.test\\.js$",
-    "transform": {
-      "^.+\\.js$": "babel-jest"
-    },
-    "transformIgnorePatterns": [
-      "/node_modules/*.+\\.js$"
-    ]
+    "postinstall": "./node_modules/.bin/gulp"
   },
   "dependencies": {
-    "gulp-replace": "^1.0.0",
-    "node-sass": "^4.12.0"
+    "gulp-replace": "^1.0.0"
   }
 }
