+++ templates/Case/index.html.twig
@@ -1,4 +1,4 @@
-{% extends 'base.html.twig' %}
+{% extends 'Layouts/base-template.html.twig' %}
 
 {% block htmlTitle %}Orders{% endblock %}
 {% block pageTitle %}
