+++ templates/Common/_timeout.html.twig
@@ -1,17 +1,9 @@
 {% set translationDomain = "popup-timeout" %}
 {% trans_default_domain translationDomain %}
 
-{% block javascripts %}
-    {{ encore_entry_script_tags('timeout') }}
-{% endblock %}
 
-{#Using data attributes to pass twig vars to javascript#}
-<div class="session-timeout-underlay"
-     data-session-expires="{{ session_expires_seconds|default(1500) }}"
-     data-session-popup-show-after="{{ session_popup_show_after|default(1200) }}"
-     data-keep-alive-url="{{ path('session-keep-alive') }}">
+<div class="session-timeout-underlay">
 </div>
-
 <div id="timeoutPopup" data-role="timeout-popup" class="session-timeout js-session-timeout">
     <div class="session-timeout-dialog">
 
@@ -31,3 +23,23 @@
 
     </div>
 </div>
+
+<script>
+
+    // instantiate object and attach events on page load
+    $().ready(function () { // $DOM READY EVENT
+
+        var sessionExpires = {{ session_expires_seconds|default(1500) }},
+            sessionShowPopupMs = {{ session_popup_show_after|default(1200) }},
+            keepAliveUrl = '{{ path('session-keep-alive') }}';
+
+        new SessionTimeoutDialog({
+            'element': $('#timeoutPopup'),
+            'sessionExpiresMs': sessionExpires * 1000,
+            'sessionPopupShowAfterMs': sessionShowPopupMs * 1000,
+            'keepSessionAliveUrl': keepAliveUrl
+        }).startCountdown();
+
+    });
+
+</script>
