+++ templates/Common/confirm.html.twig
@@ -1,4 +1,4 @@
-{% extends 'base.html.twig' %}
+{% extends 'Layouts/base-template.html.twig' %}
 {% form_theme form 'Form/govuk_frontend.html.twig' %}
 
 {% block htmlTitle %}Confirm{% endblock %}
