+++ templates/Deputy/add.html.twig
@@ -1,4 +1,4 @@
-{% extends 'base.html.twig' %}
+{% extends 'Layouts/base-template.html.twig' %}
 {% form_theme form 'Form/govuk_frontend.html.twig' %}
 {% import 'Macros/_macros.html.twig' as macro %}
 
@@ -11,11 +11,6 @@
 
 {% block validationSummary %}
     {{ macro.errorSummary(form) }}
-{% endblock %}
-
-{% block javascripts %}
-    {{ parent('app') }}
-    {{ encore_entry_script_tags('addDeputy') }}
 {% endblock %}
 
 {% block content %}
