+++ templates/Index/design.html.twig
@@ -1,4 +1,4 @@
-{% extends 'base.html.twig' %}
+{% extends 'Layouts/base-template.html.twig' %}
 
 {% block htmlTitle %}Design stuff{% endblock %}
 {% block pageTitle %}
