+++ templates/Index/index.html.twig
@@ -1,4 +1,4 @@
-{% extends 'base.html.twig' %}
+{% extends 'Layouts/base-template.html.twig' %}
 
 {% block pageTitle %}
     <h1 class="govuk-heading-xl">Homepage</h1>
