+++ templates/Report/case-report.html.twig
@@ -1,4 +1,4 @@
-{% extends 'base.html.twig' %}
+{% extends 'Layouts/base-template.html.twig' %}
 
 {% block htmlTitle %}Served Orders Report{% endblock %}
