+++ templates/User/login.html.twig
@@ -1,4 +1,4 @@
-{% extends 'base.html.twig' %}
+{% extends 'Layouts/base-template.html.twig' %}
 
 {% block htmlTitle %}Sign in{% endblock %}
 {% block pageTitle %}
@@ -33,7 +33,7 @@
 
 <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
-        <form id="login" action="{{ path('login') }}" method="post" autocomplete="on">
+        <form id="login" action="{{ path('login') }}" method="post" autocomplete="off">
 
             <div class="govuk-form-group">
                 <fieldset class="govuk-fieldset" aria-describedby="" role="group">
