+++ templates/User/password-reset-request.html.twig
@@ -1,4 +1,4 @@
-{% extends 'base.html.twig' %}
+{% extends 'Layouts/base-template.html.twig' %}
 {% form_theme form 'Form/govuk_frontend.html.twig' %}
 {% import 'Macros/_macros.html.twig' as macro %}
