+++ templates/User/password-reset-sent.html.twig
@@ -1,4 +1,4 @@
-{% extends 'base.html.twig' %}
+{% extends 'Layouts/base-template.html.twig' %}
 {% import 'Macros/_macros.html.twig' as macro %}
 
 {% block htmlTitle %}Forgotten your password - Thank you{% endblock %}
